feat: Improve type system
By using the Nominal type to tag our primitive TS types with a unique key and it's corresponding JS type, we're able to achieve a similar level of _compile time_ type safety to C. This also enables us to easily infer the return type from the Library.invoke method based on the Koffi alias we pass as the return type argument. We're still missing a bunch of types, most notably the LP* types, but those will be added as needed.